backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
pydantic==2.10.3
pydantic[email]==2.10.3
aiosqlite==0.20.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
email-validator==2.2.0
